pyproject.toml
[project]
name = "ai-document-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.77.0",
    "fpdf2>=2.8.5",
    "pydantic>=2.12.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]
